A very large overhaul
Went over each step and actually generated a model. Tested the full model and separated that into a separate R file. Went in and generated many more plots. Preformed Shaprio and BF tests to check for significance. Found a model that satisfies all of our assumptions. Finished up with some future directions for us to go.